# heavily optimize the `Value` type