This PR has many commits but only one I would say is "big". This logic is all pretty-heavily covered by tests, and I added more tests, as well as benchmarks, so hopefully you don't need to review the fiddly bit-manipulation stuff too hard.

It also leaves the `value` module in a somewhat messier state than when it started -- we have a new `ValueRef` type which duplicates some functionality of `Value` but isn't exposed in the public API; we now implement `DagLike` on `Value` itself rather than `&Value`, but using it is pretty slow compared to using it on `ValueRef`. The `Eq` and `Ord` traits are now implemented manually, but `Eq` requires a bit of computation and `Ord`, despite being implemented manually, still exposes a pretty-arbitrary ordering on values which have different types.

So it is probably worthwhile to follow up on this with a PR that overhauls the public API and tries to provide a cleaner interface.

**However**, this PR also introduces some massive speedups. It introduces new benchmarks and has their output in some commit messages. The highlights are as follows:

Prior to this PR:
```
    benches::bench_value_create_64k               ... bench:   2,565,318.90 ns/iter (+/- 40,140.60)
    benches::bench_value_create_64k_compact       ... bench: 214,621,296.60 ns/iter (+/- 23,731,924.55)
    benches::bench_value_display_64k              ... bench:  17,526,983.30 ns/iter (+/- 339,118.95)
````

With this PR:
```
    benches::bench_value_create_64k               ... bench:     875,039.00 ns/iter (+/- 5,007.70)
    benches::bench_value_create_64k_compact       ... bench:      45,258.88 ns/iter (+/- 169.22)
    benches::bench_value_display_64k              ... bench:   9,685,460.80 ns/iter (+/- 48,509.31)
```

In other words, directly creating a value which is a 64 kilobyte (not kilobit) bitstring used to take 2.5ms; displaying took 17ms; and parsing it from a bit iterator took **214** milliseconds. Now these take 0.875ms, 9ms and 0.045ms, respectively. These are a 2.8, 1.8, and 4700x speedup respectively.

The latter operation in particular, creating values from bitstrings and types, has been a major bottleneck in my fuzzer and is the motivation for this PR.