@apoelstra 2025-02-10
Drafting overnight (unless I can fix this in the next couple minutes).

Can crash it with the following regression test
```diff

diff --git a/src/value.rs b/src/value.rs
index 525b163..20d069c 100644
--- a/src/value.rs
+++ b/src/value.rs
@@ -854,4 +854,20 @@ mod tests {
     }

+    #[test]
+    fn prune_regression_1() {
+        // Found this when fuzzing Elements; unsure how to reduce it further.
+        let nontrivial_sum = Value::product(
+            Value::right(Final::two_two_n(4), Value::u16(0)),
+            Value::u8(0),
+        );
+        // Formatting should succeed and have no effect.
+        let _ = format!("{nontrivial_sum}");
+        // Pruning should succeed and have no effect.
+        assert_eq!(
+            nontrivial_sum.prune(nontrivial_sum.ty()),
+            Some(nontrivial_sum)
+        );
+    }
+
     #[test]
     fn prune() {
```