@apoelstra 2025-02-10
Addressed every comment labelled on cbbe9b5. I believe all the rest are followups or non-actionable.